debezium/dbz#2379 Move claim-check serializer into dedicated package
